// src/api.ts
export class Deployment {
    'service': string;
    'roles': { [key: string]: InlineResponse2001Roles; };
}
// ...
export class InlineResponse2001Instances {
    'id': string;
    'privateIp': string;
    'publicIp': string;
    'arrangement': InlineResponse2001Arrangement;
}

export class InlineResponse2001Roles {
    'instances': { [key: string]: InlineResponse2001Instances; };
}